rag_with_python/real_estate_rag_milvus/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymilvus==2.3.1
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
pypdf2==3.0.1
numpy==1.26.1
pydantic==2.5.1
pydantic-settings==2.0.3
openai==1.3.5